add the ability to specify a custom OCR besides the one natively supported.
